Add panels and queries for Octavia, Swift, and Manila to current availability dashboard
Added panels and queries for these three openstack components. The layout of the dashboard has also been updated as well.
The uid for the dashboard has been updated so that it is clearer what the dashboard is based on its uid.
Panels for Placement and the web interface have been removed as there is no test for these currently.
Panels for storage and network overview have also been removed as no query had been set up for them (may be added in the future).