MAINT: Remove 30d cap from panels
The panels had a 30d cap set manually on the graphs. This was due to the
way the data used to be sent to the datasource but has since been
improved. Removing this cap allows us to see graph trends up to
whatever time period the page is set to.